Add support for automating labels
This commit adds support for automating the labels across all
organizations in the unified collective.  The GitHub label defaults are
broad and not very useful.  These labels are specific to our needs.

This comes with a defined structure of labels that are useful to our
collective.  They come in groups of area, platform, semver, status,
type, and there are a couple of miscellaneous labels.

The tooling now warns for unexpected labels, adds missing labels, fixes
incorrect names, colors, and descriptions, replaces old labels with new
on existing issues and pull requests, and removes legacy labels.

One cool extra is that the structured labels support automating color
scaling for groups!

Closes GH-4.

Reviewed-by: Christian Murphy <christian.murphy.42@gmail.com>